this doesn't need to be 64-bit (i think), i believe it is a relative position

+++ source/CallTip.h
@@ -2,8 +2,6 @@
 #ifndef CALL_TIP_H_
 #define CALL_TIP_H_
 
-#include <cstdint>
-
 enum class TipHAlignMode  { Left, Center, Right };
 enum class TipVAlignMode  { Above, Below };
 enum class TipAlignStrict { Sloppy, Strict };
@@ -11,7 +9,7 @@ enum class TipAlignStrict { Sloppy, Strict };
 struct CallTip {
     int ID;                   //  ID of displayed calltip.  Equals zero if none is displayed.
     bool anchored;            //  Is it anchored to a position
-    int64_t pos;              //  Position tip is anchored to
+    int pos;                  //  Position tip is anchored to
     TipHAlignMode hAlign;     //  horizontal alignment
     TipVAlignMode vAlign;     //  vertical alignment
     TipAlignStrict alignMode; //  Strict or sloppy alignment
